refactor(auth): extract rate_limit seam
Move the login-rate-limit state + helpers to a dedicated module:

- _login_attempts, _LOGIN_MAX_ATTEMPTS, _LOGIN_WINDOW_SECONDS
- _client_ip, check_login_rate_limit, record_failed_login,
  clear_login_attempts
- new reset_login_attempts() helper for cross-seam callers
  (reset_auth_caches composes over it so no other seam reaches
   into the rate-limit dict directly)

auth/__init__.py re-exports every name that was public or poked
by tests (conftest.py clears _login_attempts between tests; admin
tests assert _login_attempts == {}); __all__ lists the re-exports
explicitly so mypy's attr-defined check stays green without
touching any downstream caller.
